apps/api/core/serializers.py
from rest_framework import serializers


class HealthCheckSerializer(serializers.Serializer):
    status = serializers.CharField()
    message = serializers.CharField()
    version = serializers.CharField()