apps/api/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
Pillow==10.3.0
boto3==1.34.109
django-storages==1.14.3
django-filter==24.2
drf-spectacular==0.27.2
python-dotenv==1.0.1